Time: 0 ms (100.00%) | Memory: 63 MB (10.93%) - LeetSync

+++ 11-container-with-most-water/container-with-most-water.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     int maxArea(vector<int>& height) {
         int l = 0;
-        long long r = height.size()-1;
+        int r = height.size()-1;
         long long maxArea = 0;
         while(l < r){
-            long long ar = (r-l) * min(height[l], height[r]);
+            long long ar = (long long)(r-l) * min(height[l], height[r]);
             maxArea = max(maxArea, ar);
             if(height[l] < height[r]){
                 l++;
